% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{aspwbInput}
\alias{aspwbInput}
\alias{aspwb_day}
\alias{aspwb}
\title{Simulation in agricultural areas}
\usage{
aspwbInput(crop_factor, control, soil)

aspwb_day(
  x,
  date,
  meteovec,
  latitude,
  elevation,
  slope,
  aspect,
  runon = 0,
  modifyInput = TRUE
)

aspwb(
  x,
  meteo,
  latitude,
  elevation = NA_real_,
  slope = NA_real_,
  aspect = NA_real_
)
}
\arguments{
\item{crop_factor}{Agriculture crop factor.}

\item{control}{A list with default control parameters (see \code{\link{defaultControl}}).}

\item{soil}{An object of class \code{\link{soil}}.}

\item{x}{An object of class \code{\link{aspwbInput}}.}

\item{date}{Date as string "yyyy-mm-dd".}

\item{meteovec}{A named numerical vector with weather data. See variable names in parameter \code{meteo} of \code{\link{spwb}}.}

\item{latitude}{Latitude (in degrees).}

\item{elevation, slope, aspect}{Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North).}

\item{runon}{Surface water amount running on the target area from upslope (in mm).}

\item{modifyInput}{Boolean flag to indicate that the input \code{x} object is allowed to be modified during the simulation.}

\item{meteo}{A data frame with daily meteorological data series (see \code{\link{spwb}}).}
}
\description{
Function \code{aspwb_day} performs water balance for a single day in an agriculture location.
}
\examples{

control <- defaultControl()
examplesoil <- soil(defaultSoilParams())
x <- aspwbInput(0.75, control, examplesoil)

#Call simulation function
S <- aspwb(x, examplemeteo, latitude = 41.82592, elevation = 100)

}
\seealso{
\code{\link{spwbInput}}, \code{\link{spwb}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
