% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{communication}
\alias{communication}
\alias{.copy_model_output}
\title{Internal communication}
\usage{
.copy_model_output(internalCommunication, x, model)
}
\arguments{
\item{internalCommunication}{List for internal communication.}

\item{x}{An object of class \code{\link{spwbInput}} or \code{\link{growthInput}}.}

\item{model}{String for model, either "spwb" or "growth".}

\item{date}{Date as string "yyyy-mm-dd".}

\item{meteovec}{A named numerical vector with weather data. See variable names in parameter \code{meteo} of \code{\link{spwb}}.}

\item{latitude}{Latitude (in degrees).}

\item{elevation, slope, aspect}{Elevation above sea level (in m), slope (in degrees) and aspect (in degrees from North).}

\item{runon}{Surface water amount running on the target area from upslope (in mm).}

\item{lateralFlows}{Lateral source/sink terms for each soil layer (interflow/to from adjacent locations) as mm/day.}

\item{waterTableDepth}{Water table depth (in mm). When not missing, capillarity rise will be allowed if lower than total soil depth.}

\item{modifyInput}{Boolean flag to indicate that the input \code{x} object is allowed to be modified during the simulation.}
}
\description{
Functions for internal communication. Not to be called by users.
}
\keyword{internal}
