% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomposition_annualLitterDecompositionRate}
\alias{decomposition_annualLitterDecompositionRate}
\alias{decomposition_litterMetabolicFraction}
\alias{decomposition_pHEffect}
\alias{decomposition_moistureEffect}
\alias{decomposition_temperatureEffect}
\title{Low-level decomposition functions}
\usage{
decomposition_annualLitterDecompositionRate(AET, lignin)

decomposition_litterMetabolicFraction(ligninPercent, Nmass)

decomposition_pHEffect(x, pool)

decomposition_moistureEffect(sand, clay, soilMoisture)

decomposition_temperatureEffect(soilTemperature)
}
\arguments{
\item{AET}{Actual evapotranspiration (mm)}

\item{lignin}{Lignin percent}

\item{ligninPercent}{lignin content (\% of dry)}

\item{Nmass}{nitrogen content (mg N / g dry)}

\item{x}{Soil water pH (0-14)}

\item{pool}{String indicating the decomposition pool}

\item{sand, clay}{Soil texture values in percent volume.}

\item{soilMoisture}{Soil moisture content, relative to saturation.}

\item{soilTemperature}{Soil temperature (in Celsius).}
}
\value{
Functions \code{decomposition_moistureEffect}, \code{decomposition_pHEffect} and \code{decomposition_temperatureEffect} return
a scalar value representing a factor that should modify a decomposition rate. Function \code{decomposition_annualLitterDecompositionRate}
directly returns a scalar value with the annual decomposition rate (yr-1). Function \code{decomposition_litterMetabolicFraction} returns
a scalar with the fraction of litter that corresponds to metabolic carbon.
}
\description{
Functions related to litter and soil carbon decomposition processes
}
\details{
Function \code{decomposition_moistureEffect} follows Kelly et al. (2000)
}
\references{
Bonan, G. (2019). Climate change and terrestrial ecosystem modeling. Cambridge University Press, Cambridge, UK.

Meentemeyer (1978)
Kelly et al (2000)
}
\seealso{
\code{\link{decomposition_DAYCENT}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{internal}
