% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomposition_DAYCENT}
\alias{decomposition_DAYCENT}
\title{DAYCENT decomposition}
\usage{
decomposition_DAYCENT(
  snags,
  litter,
  SOC,
  paramsLitterDecomposition,
  baseAnnualRates,
  annualTurnoverRate,
  environmentalConditions,
  litterProduction,
  sand,
  clay,
  soilPH,
  soilO2 = 1,
  cultfac = 1,
  tstep = 1,
  balanceCheck = FALSE
)
}
\arguments{
\item{snags}{A data frame with initial dead standing (snag) cohort information (see \code{\link{growthInput}}).}

\item{litter}{A data frame with initial aboveground and belowground structural carbon pools corresponding to plant cohorts, in g C/m2  (see \code{\link{growthInput}}).}

\item{SOC}{A named numeric vector with initial metabolic, active, slow and passive carbon pools for surface and soil, in g C/m2  (see \code{\link{growthInput}}).}

\item{paramsLitterDecomposition}{A data frame of species-specific litter decomposition parameters (see \code{\link{growthInput}}).}

\item{baseAnnualRates}{A named vector of annual decomposition rates, in yr-1 (see \code{\link{defaultControl}}).}

\item{annualTurnoverRate}{Annual turnover rate, in yr-1  (see \code{\link{defaultControl}}).}

\item{environmentalConditions}{A data frame containing environmental conditions for each time step to simulate:
\itemize{
\item{\code{AirTempperature}: Mean air temperature (in Celsius).}
\item{\code{AirRelativeHumidity}: Mean relative humidity (percent).}
\item{\code{SoilTemperature}: Mean soil temperature (in Celsius).}
\item{\code{SoilMoisture}: Mean soil moisture, relative to saturation (0-1).}
}}

\item{litterProduction}{A data frame containing litter inputs corresponding to time steps:
\itemize{
\item{\code{Step}: Integer indicating the time step where litter is generated, corresponding to a row in \code{environmentalConditions}.}
\item{\code{Species}: Mean relative humidity (percent).}
\item{\code{Leaves}: Leaf litter production (g C·m-2).}
\item{\code{Twigs}: Twig litter production (g C·m-2).}
\item{\code{SmallBranches}: Small branch litter production (g C·m-2).}
\item{\code{LargeWood}: Large wood litter production (g C·m-2).}
\item{\code{CoarseRoots}: Coarse root litter production (g C·m-2).}
\item{\code{FineRoots}: Fine root litter production (g C·m-2).}
}}

\item{sand, clay}{Soil texture (sand and sand) in percent volume (percent).}

\item{soilPH}{Soil pH (0-14).}

\item{soilO2}{Soil oxygen factor (0-1).}

\item{cultfac}{Cultivation factor (0-1).}

\item{tstep}{Time step in days. By default, one day. For annual time steps, use \code{tstep = 365.25}.}

\item{balanceCheck}{Flag to print carbon balance check (for debugging).}
}
\value{
A list with two elements:
\itemize{
\item{\code{"HeterotrophicRespiration"}: A numeric matrix with the heterotrophic respiration (g C·m-2) corresponding to the decomposition in each time step corresponding to different carbon pools:
\itemize{
\item{\code{"SurfaceSnags"}: Respiration from decomposing snags (i.e. small branches and large wood).}
\item{\code{"SurfaceLitter"}: Respiration from decomposing litter in the surface (i.e. leaves, twigs, small branches and large wood).}
\item{\code{"SoilLitter"}: Respiration from decomposing litter belowground (i.e. coarse roots and fine roots).}
\item{\code{"SurfaceMetabolic"}: Respiration from surface metabolic pool.}
\item{\code{"SoilMetabolic"}: Respiration from soil metabolic pool.}
\item{\code{"SurfaceActive"}: Respiration from surface active pool.}
\item{\code{"SoilActive"}: Respiration from soil active pool.}
\item{\code{"SurfaceSlow"}: Respiration from surface slow pool.}
\item{\code{"SoilSlow"}: Respiration from soil slow pool.}
\item{\code{"SoilPassive"}: Respiration from soil passive pool.}
}
}
\item{\code{"DecompositionPools"}: A numeric matrix with the carbon mass of different carbon pools, all in g C · m-2:
\itemize{
\item{\code{"SmallBranchSnags"}: Carbon of (standing) small dead branches (diameters between 0.635 cm and 7.5 cm; 10h and 100h fuels).}
\item{\code{"LargeWoodSnags"}: Carbon of (standing) large dead wood (diameters larger than 7.5 cm; 1000h and 10000h fuels).}
\item{\code{"LeafLitter"}: Carbon of leaf litter.}
\item{\code{"TwigLitter"}: Carbon of twig litter.}
\item{\code{"SmallBranchLitter"}: Carbon of (downed) small dead branches (diameters between 0.635 cm and 7.5 cm; 10h and 100h fuels).}
\item{\code{"LargeWoodLitter"}: Carbon of (downed) large dead wood (diameters larger than 7.5 cm; 1000h and 10000h).}
\item{\code{"CoarseRootLitter"}: Carbon of dead coarse roots.}
\item{\code{"FineRootLitter"}: Carbon of fine root litter.}
\item{\code{"SurfaceMetabolic"}: Metabolic carbon in the surface.}
\item{\code{"SoilMetabolic"}: Metabolic carbon in the soil.}
\item{\code{"SurfaceActive"}: Active decomposition carbon pool in the surface.}
\item{\code{"SoilActive"}: Active decomposition carbon pool in the soil.}
\item{\code{"SurfaceSlow"}: Slow decomposition carbon pool in the surface.}
\item{\code{"SoilSlow"}: Slow decomposition carbon pool in the soil.}
\item{\code{"SoilPassive"}: Passive decomposition carbon pool in the soil.}
}
}
}
}
\description{
Function implementing a modification of the DAYCENT carbon decomposition model (Parton et al. 1988, 1993, 1998), inspired by the description given in Chapter 18 of Bonan (2019).
}
\details{
Species names must match between inputs \code{paramsLitterDecomposition}, \code{litter} and \code{litterProduction}.

\emph{IMPORTANT NOTE}: Decomposition functions modify the input data (i.e. \code{snags}, \code{litter} and/or \code{SOC}) according to decomposition rates and carbon transfer rates. When used as part of \code{\link{growth}} simulations,
soil physical and chemical parameters correspond to the uppermost soil layer.
}
\references{
Bonan, G. (2019). Climate change and terrestrial ecosystem modeling. Cambridge University Press, Cambridge, UK.

Parton WJ, Steward JWB, Cole CV (1988). Dynamics of C, N, P and S in grassland soils: a model. Biogeochemistry 5: 109-131.

Parton WJ, Scurlock JMO, Ojima DS, Gilmanov TG, Scoles RJ et al. (1993). Observations and modeling of biomass and soil organic matter dynamics for the grassland biome worldwide. Global Biogeochemical Cycles 7: 785-809.

Parton WJ, Hartman M, Ojima DS, Schimel D (1998). DAYCENT and its land surface submodel: Description and testing. Global and Planetary Change, 19: 35-48.
}
\seealso{
\code{\link{decomposition_temperatureEffect}}, \code{\link{growthInput}}, \code{\link{growth}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF

Roberto Molowny-Horas, CREAF

Inés Delsman Valderrama, CREAF
}
\keyword{internal}
