% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomposition_moistureEffect}
\alias{decomposition_moistureEffect}
\title{Moisture effect}
\usage{
decomposition_moistureEffect(sand, clay, soilMoisture)
}
\arguments{
\item{sand, }{clay texture values in percent.}

\item{soilMoisture}{Soil moisture content, relative to saturation.}
}
\description{
Moisture factor in CENTURY (from Kelly et al 2000)
}
