\encoding{UTF-8}
\name{exampleforest}
\Rdversion{1.1}
\alias{exampleforest}
\docType{data}
\title{Example forest stand}
\description{
  Data set for illustration of model behaviour. Includes a description of the plant cohorts of a forest stand.
}
\usage{data(exampleforest)}
\format{
 An object of class \code{\link{forest}} containing the description of the tree, sapling and shrub cohorts of a forest patch as well as the seed bank and the size of the patch.
}
\source{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. Dirección General de Conservación de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\seealso{ \code{\link{forest}}, \code{\link{swb}}, \code{\link{forest2swbInput}}}
\examples{
data(exampleforest)
}
\keyword{datasets}
