% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{exampleobs}
\alias{exampleobs}
\title{Example observed data}
\format{
A data frame containing daily 'observed' values for year 2001:
\describe{
\item{\code{dates}}{Measurement dates.}
\item{\code{SWC}}{Soil moisture content (in m3/m3).}
\item{\code{ETR}}{Total evapotranspiration (mm).}
\item{\code{E_T1_158}}{Transpiration of Pinus halepensis cohort 'T1_158' (L/m2 of leaf area).}
\item{\code{E_T2_179}}{Transpiration of Quercus ilex cohort 'T2_179' (L/m2 of leaf area).}
\item{\code{FMC_T1_158}}{Fuel moisture content of Pinus halepensis cohort 'T1_158'  (in percent).}
\item{\code{FMC_T2_179}}{Fuel moisture content of Quercus ilex cohort 'T2_179' (in percent).}
\item{\code{BAI_T1_158}}{Basal area increment for Pinus halepensis cohort 'T1_158'  (in cm2).}
\item{\code{BAI_T2_179}}{Basal area increment for Quercus ilex cohort 'T2_179' (in cm2).}
\item{\code{DI_T1_158}}{Diameter increment for Pinus halepensis cohort 'T1_158'  (in cm).}
\item{\code{DI_T2_179}}{Diameter increment for Quercus ilex cohort 'T2_179' (in cm).}
}
}
\source{
This data set was actually created by running a simulation and adding some gaussian error to the outputs.
}
\description{
Example (fake) data set of variables measured in a plot.
}
\examples{
data(exampleobs)
}
\seealso{
\code{\link{evaluation}}
}
\keyword{datasets}
