% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{forest}
\alias{forest}
\alias{exampleforest}
\alias{exampleforest2}
\title{Description of a forest stand.}
\format{
An object of class \code{forest} contains the description of the woody (tree or shrub) cohorts of a forest patch.
It has the following structure (see additional columns in details):
\itemize{
\item{\code{treeData}: A data frame of tree cohorts (in rows) and the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name or a non-negative integer for tree species identity (i.e., 0,1,2,...) matching SpParams.}
\item{\code{Height}: Total tree height (in cm).}
\item{\code{DBH}: Tree diameter at breast height (in cm).}
\item{\code{N}: Density (number of individuals/hectare) that the measured tree represents.}
\item{\code{Z50}: Depth (in mm) corresponding to 50\% of fine roots.}
\item{\code{Z95}: Depth (in mm) corresponding to 95\% of fine roots.}
}
}
\item{\code{shrubData}: A data frame of shrub cohorts (in rows) and the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name or a non-negative integer for shrub species identity (i.e., 0,1,2,...) matching SpParams.}
\item{\code{Height}: Average total height of plants (in cm).}
\item{\code{Cover}: Percent cover.}
\item{\code{Z50}: Depth (in mm) corresponding to 50\% of fine roots.}
\item{\code{Z95}: Depth (in mm) corresponding to 95\% of fine roots.}
}
}
}
Additionally, it can contain information about other elements: the herbaceous layer, seedling/sapling bank, seed bank, snags, litter or soil carbon:
\itemize{
\item{\code{herbData}: A data frame of herbaceous cohorts (in rows) and the following columns (see an alternative way of defining the herbaceous layer in details):
\itemize{
\item{\code{Species}: String with species (taxon) name or a non-negative integer for herbaceous species identity (i.e., 0,1,2,...) matching SpParams.}
\item{\code{Height}: Average total height of plants (in cm).}
\item{\code{Cover}: Percent cover.}
\item{\code{Z50}: Depth (in mm) corresponding to 50\% of fine roots.}
\item{\code{Z95}: Depth (in mm) corresponding to 95\% of fine roots.}
}
}
\item{\code{seedlingBank}: An optional data frame containing seedling/sapling information with the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name or a non-negative integer for seedling/sapling species identity (i.e., 0,1,2,...) matching SpParams.}
\item{\code{Percent}: Amount of seedling in relation to full seedling bank (in \%).}
\item{\code{Age}: A numeric vector indicating age of seedlings/saplings in years.}
\item{\code{Z50}: Depth (in mm) corresponding to 50\% of fine roots.}
\item{\code{Z95}: Depth (in mm) corresponding to 95\% of fine roots.}
}
}
\item{\code{seedBank}: An optional data frame containing seed bank information with the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name or a non-negative integer for seed species identity (i.e., 0,1,2,...) matching SpParams.}
\item{\code{Percent}: Amount of seeds in relation to full seed bank (in \%).}
}
}
\item{\code{snagData}: An optional data frame containing information for snag cohorts, with the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name.}
\item{\code{DBH}: Diameter at breast height (in cm), for tree snags only.}
\item{\code{Height}: Height of the snag cohort (tree or shrub) (in cm).}
\item{\code{DeathAge}: Years passed since plant death (in yrs).}
\item{\code{SmallBranches}: Mass of (standing) small dead branches (in g C/m2).}
\item{\code{LargeWood}: Mass of (standing) large dead wood (in g C/m2).}
}
}
\item{\code{litterData}: An optional data frame containing litter information with the following columns:
\itemize{
\item{\code{Species}: String with species (taxon) name.}
\item{\code{Leaves}: Mass of leaf litter (in g C/m2)}
\item{\code{Twigs}: Mass of twig litter (in g C/m2)}
\item{\code{SmallBranches}: Mass of (downed) small dead branches (in g C/m2)}
\item{\code{LargeWood}: Mass of (downed) large dead wood (in g C/m2)}
\item{\code{CoarseRoots}: Mass of dead coarse roots (in g C/m2)}
\item{\code{FineRoots}: Mass of fine root litter (in g C/m2)}
}
}
\item{\code{SOCData}: An optional named numeric vector containing CENTURY soil organic carbon pools (normally from previous simulations):
\itemize{
\item{\code{SurfaceMetabolic}: Metabolic carbon in the surface (in g C/m2).}
\item{\code{SoilMetabolic}: Metabolic carbon in the soil (in g C/m2)}
\item{\code{SurfaceActive}: Active decomposition carbon pool in the surface (in g C/m2)}
\item{\code{SoilActive}: Active decomposition carbon pool in the soil (in g C/m2)}
\item{\code{SurfaceSlow}: Slow decomposition carbon pool in the surface (in g C/m2)}
\item{\code{SoilSlow}: Slow decomposition carbon pool in the soil (in g C/m2)}
\item{\code{SoilPassive}: Passive decomposition carbon pool in the soil (in g C/m2)}
}
}
}
}
\source{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. \enc{Dirección}{Direccion} General de \enc{Conservación}{Conservacion} de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\description{
\code{exampleforest} is an example of forest stand description, whereas \code{exampleforest2} is an alternative forest description where leaf area index and crown ratio are supplied instead of structural (density, DBH and cover) parameters.
}
\details{
The structure presented above for \code{forest} objects corresponds to the main (required and optional) data elements and their required columns.
A \code{forest} object can contain additional information when this is available. Data frames \code{treeData},
\code{shrubData} and \code{herbData} can contain additional columns (\code{Z100} is also commonly used for data frame \code{seedlingBank}):
\itemize{
\item{\code{Z100}: Depth (in mm) corresponding to 100\% of fine roots (to specify a truncated root distribution).}
\item{\code{Age}: A numeric vector indicating age of cohorts in years. Used to track cohort age in simulations with \code{\link{fordyn}}.}
\item{\code{LAI}: Leaf area index (m2/m2).}
\item{\code{FoliarBiomass}: Standing dry biomass of leaves (kg/m2).}
\item{\code{FuelLoading}: Fine fuel loading (kg/m2).}
\item{\code{CrownRatio}: The ratio between crown length and total height (between 0 and 1)}
\item{\code{ObsID}: A string identifying plant cohorts at the stage of forest sampling. Used to track the fate of particular plant cohorts through simulations.
In \code{\link{fordyn}}, the use of these identifiers can be combined with the control option \code{keepCohortsWithObsID} so that these cohorts are not merged
or removed during simulations.}
}

Leaf area index (\code{LAI}), foliar biomass (\code{FoliarBiomass}), fuel loading (\code{FuelLoading}) and \code{CrownRatio} are used to override allometry-based estimates of those variables
when initializing inputs for functions \code{\link{spwb}} or \code{\link{spwb_day}}, but they will have no effect in simulations of \code{\link{growth}}, \code{\link{growth_day}} or \code{\link{fordyn}}.
Note that leaf area index, foliar biomass and fuel loading are related entities, and they are treated as such in medfate. Therefore, users are expected to supply
one or the other, and not all of them at the same time.

Instead of defining \code{herbData}, the herbaceous layer can be defined collectively (avoiding the need to specify species parameters for herbaceous species),
using the following list items:
\itemize{
\item{\code{herbCover}: Percent cover of the herb layer.}
\item{\code{herbHeight}: Mean height (in cm) of the herb layer.}
\item{\code{herbLAI}: Leaf area index (m2/m2) of the herb layer.}
\item{\code{herbFoliarBiomass}: Dry biomass (kg/m2) of the herb layer.}
\item{\code{herbFuelLoading}: Fine fuel loading (kg/m2) of the herb layer.}
}
}
\examples{
data(exampleforest)
exampleforest

data(exampleforest2)
exampleforest2
}
\seealso{
\code{\link{summary.forest}}, \code{\link{emptyforest}}, \code{\link{plot.forest}}, \code{\link{spwbInput}}
}
\keyword{data}
