\encoding{UTF-8}
\name{Forest manipulation}
\alias{forest_mergeTrees}

\title{Forest utility functions}
\description{
Functions to manipulate a \code{\link{forest}} object.
}
\usage{
forest_mergeTrees(x)
}
\arguments{
  \item{x}{An object of class \code{\link{forest}}.}
}
\value{
Depending on the function:
\itemize{
\item{\code{forest_mergeTrees}: Another \code{\link{forest}} object with merged trees.}
}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CTFC
}
\seealso{
 \code{\link{spwb}}, \code{\link{forest}}, \code{\link{summary.forest}}
}