\name{getSpParamsDefinition}
\alias{getSpParamsDefinition}
\title{
Species parameter definition
}
\description{
Species parameter definition table
}
\usage{
getSpParamsDefinition()
}
\value{
A data frame giving the definition of species parameter values
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\seealso{
\code{\link{SpParamsMED}}
}
