\name{getSpParamsNames}
\alias{getSpParamsNames}
\title{
Species parameter names
}
\description{
Vector of species parameter names
}
\usage{
getSpParamsNames()
}
\details{
Parameter definitions are found in \code{\link{SpParamsMED}}
}
\value{
A string vector
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\seealso{
\code{\link{SpParamsMED}}
}
