% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{herb_values}
\alias{herb_values}
\alias{herb_foliarBiomass}
\alias{herb_fuel}
\alias{herb_LAI}
\title{Herbaceous description functions}
\usage{
herb_foliarBiomass(herbCover, herbHeight)

herb_fuel(herbCover, herbHeight)

herb_LAI(herbCover, herbHeight)
}
\arguments{
\item{herbCover}{Percent cover of the herb layer.}

\item{herbHeight}{Mean height (in cm) of the herb layer.}
}
\value{
A single scalar:
\itemize{
  \item{\code{herb_foliarBiomass}: Biomass of leaves (in kg/m2).}
  \item{\code{herb_fuel}: Fine fuel load (in kg/m2).}
  \item{\code{herb_LAI}: Leaf area index (m2/m2).}
}
}
\description{
Functions to calculate attributes of the herbaceous component of a \code{\link{forest}} object 
via its attributes herbCover and herbHeight.
}
\seealso{
\code{\link{spwb}}, \code{\link{forest}}, \code{\link{plant_basalArea}}, \code{\link{summary.forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
