% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hydrology_infiltration}
\alias{hydrology_infiltration}
\alias{hydrology_infiltrationBoughton}
\alias{hydrology_infiltrationGreenAmpt}
\alias{hydrology_infiltrationRepartition}
\alias{hydrology_infiltrationAmount}
\title{Soil infiltration}
\usage{
hydrology_infiltrationBoughton(input, Ssoil)

hydrology_infiltrationGreenAmpt(t, psi_w, Ksat, theta_sat, theta_dry)

hydrology_infiltrationRepartition(I, dVec, macro, a = -0.005, b = 3)

hydrology_infiltrationAmount(
  rainfallInput,
  r,
  soil,
  soilFunctions,
  model = "Green-Ampt"
)
}
\arguments{
\item{input}{A numeric vector of (daily) water input (in mm of water).}

\item{Ssoil}{Soil water storage capacity (can be referred to topsoil) (in mm of water).}

\item{t}{Time of the infiltration event}

\item{psi_w}{Matric potential at the wetting front}

\item{Ksat}{hydraulic conductivity at saturation}

\item{theta_sat}{volumetric content at saturation}

\item{theta_dry}{volumetric content at the dry side of the wetting front}

\item{I}{Soil infiltration (in mm of water).}

\item{dVec}{Width of soil layers (in mm).}

\item{macro}{Macroporosity of soil layers (in \%).}

\item{a, b}{Parameters of the extinction function used for water infiltration.}

\item{rainfallInput}{Water from the rainfall event reaching the soil surface (mm)}

\item{r}{rainfall rate (mm/h)}

\item{model}{Infiltration model, either "Green-Ampt" or "Boughton"}
}
\value{
Function \code{hydrology_infiltrationBoughton} a vector of the same length as \code{input} containing the daily amount of water that infiltrates into the soil (in mm of water). 

Function \code{hydrology_infiltrationRepartition} estimates the amount of infiltrated water that reaches each soil layer.
}
\description{
Soil infiltration functions:
\itemize{
  \item{Function \code{hydrology_infiltrationBoughton} calculates the amount of water that infiltrates into the topsoil, according to the USDA SCS curve number method (Boughton 1989).}
  \item{Function \code{hydrology_infiltrationRepartition} distributes infiltration among soil layers depending on macroporosity.}
}
}
\examples{
SoilDepth = c(200,400,800,1200,1500)

#TOPSOIL LAYERS
d1 = pmin(SoilDepth, 300) #<300
#SUBSOIL LAYERS
d2 = pmax(0, pmin(SoilDepth-300,1200)) #300-1500 mm
#ROCK LAYER
d3 = 4000-(d1+d2) #From SoilDepth down to 4.0 m

TS_clay = 15
TS_sand = 25
SS_clay = 15
SS_sand = 25
RL_clay = 15
RL_sand = 25
TS_gravel = 20
SS_gravel = 40
RL_gravel = 95

Theta_FC1=soil_psi2thetaSX(TS_clay, TS_sand, -33) #in m3/m3
Theta_FC2=soil_psi2thetaSX(SS_clay, SS_sand, -33) #in m3/m3
Theta_FC3=soil_psi2thetaSX(RL_clay, RL_sand, -33) #in m3/m3
pcTS_gravel = 1-(TS_gravel/100)
pcSS_gravel = 1-(SS_gravel/100)
pcRL_gravel = 1-(RL_gravel/100)
MaxVol1 = (d1*Theta_FC1*pcTS_gravel)
MaxVol2 = (d2*Theta_FC2*pcSS_gravel)
MaxVol3 = (d3*Theta_FC3*pcRL_gravel)
V = MaxVol1+MaxVol2+MaxVol3

par(mar=c(5,5,1,1), mfrow=c(1,2))
NP = seq(0,60, by=1)
plot(NP,hydrology_infiltrationBoughton(NP, V[1]), type="l", xlim=c(0,60), ylim=c(0,60), 
     ylab="Infiltration (mm)", xlab="Net rainfall (mm)", frame=FALSE)
lines(NP,hydrology_infiltrationBoughton(NP, V[2]), lty=2)
lines(NP,hydrology_infiltrationBoughton(NP, V[3]), lty=3)
lines(NP,hydrology_infiltrationBoughton(NP, V[4]), lty=4)
lines(NP,hydrology_infiltrationBoughton(NP, V[5]), lty=5)
legend("topleft", bty="n", lty=1:5, 
       legend=c(paste("d =", SoilDepth, "Vsoil =",round(V),"mm")))
plot(NP,NP-hydrology_infiltrationBoughton(NP, V[1]), type="l", xlim=c(0,60), ylim=c(0,60), 
     ylab="Runoff (mm)", xlab="Net rainfall (mm)", frame=FALSE)
lines(NP,NP-hydrology_infiltrationBoughton(NP, V[2]), lty=2)
lines(NP,NP-hydrology_infiltrationBoughton(NP, V[3]), lty=3)
lines(NP,NP-hydrology_infiltrationBoughton(NP, V[4]), lty=4)
lines(NP,NP-hydrology_infiltrationBoughton(NP, V[5]), lty=5)
legend("topleft", bty="n", lty=1:5, 
       legend=c(paste("d =", SoilDepth,"Vsoil =",round(V),"mm")))

}
\references{
Boughton (1989). A review of the USDA SCS curve number method. - Australian Journal of Soil Research 27: 511-523.
}
\seealso{
\code{\link{spwb}}, \code{\link{hydrology_soilWaterInputs}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
