% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hydrology_soilFlows}
\alias{hydrology_soilFlows}
\title{Soil flows}
\usage{
hydrology_soilFlows(
  soil,
  sourceSink,
  nsteps = 24L,
  freeDrainage = TRUE,
  modifySoil = TRUE
)
}
\arguments{
\item{soil}{Object of class \code{\link{soil}}.}

\item{sourceSink}{Source/sink term for each soil layer (from snowmelt, soil evaporation or plant transpiration/redistribution)
as mm/day.}

\item{nsteps}{Number of time steps per day}

\item{modifySoil}{Boolean flag to indicate that the input \code{soil} object should be modified during the simulation.}
}
\value{
Returns a named vector with two elements:
  \itemize{
    \item{\code{deep_drainage}: the water draining from the bottom layer.}
    \item{\code{saturation_excess}: surface runoff generated via saturation excess.}
  }
}
\description{
Function \code{hydrology_soilFlows} estimates water movement within the soil according to Richards equation.
}
\seealso{
\code{\link{spwb}}, \code{\link{hydrology_soilWaterInputs}}, \code{\link{hydrology_infiltration}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
