% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hydrology_snowMelt}
\alias{hydrology_snowMelt}
\alias{hydrology_verticalInputs}
\alias{hydrology_soilWaterInputs}
\title{Soil vertical inputs}
\usage{
hydrology_snowMelt(tday, rad, LgroundSWR, elevation)

hydrology_soilWaterInputs(
  soil,
  soilFunctions,
  prec,
  rainfallIntensity,
  pet,
  tday,
  rad,
  elevation,
  Cm,
  LgroundPAR,
  LgroundSWR,
  runon = 0,
  snowpack = TRUE,
  modifySoil = TRUE
)
}
\arguments{
\item{tday}{Average day temperature (ºC).}

\item{rad}{Solar radiation (in MJ/m2/day).}

\item{LgroundSWR}{Percentage of short-wave radiation (SWR) reaching the ground.}

\item{elevation}{Altitude above sea level (m).}

\item{soil}{A list containing the description of the soil (see \code{\link{soil}}).}

\item{soilFunctions}{Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).}

\item{prec}{Precipitation for a given day (mm)}

\item{Cm}{Canopy water storage capacity.}

\item{LgroundPAR}{Percentage of photosynthetically-active radiation (PAR) reaching the ground.}

\item{runon}{Surface water amount running on the target area from upslope (in mm).}

\item{snowpack}{Boolean flag to indicate the simulation of snow accumulation and melting.}

\item{modifySoil}{Boolean flag to indicate that the input \code{soil} object should be modified during the simulation.}

\item{er}{The ratio of evaporation rate to rainfall rate.}
}
\value{
Function \code{hydrology_soilWaterInputs} returns a named vector with the following elements, all in mm:
\item{Rain}{Precipitation as rainfall.}
\item{Snow}{Precipitation as snow.}
\item{Interception}{Rainfall water intercepted by the canopy and evaporated.}
\item{NetRain}{Rainfall reaching the ground (throughfall).}
\item{Snowmelt}{Snow melted during the day, and added to the water infiltrated.}
\item{Runon}{Surface water amount running on the target area from upslope.}
\item{RainfallInput}{Rainfall input, including runon and net rain.}
\item{TotalInput}{Total soil input, including runon, snowmelt and net rain.}
}
\description{
High-level functions to define water inputs into the soil:
 
\itemize{
  \item{Function \code{hydrology_soilWaterInputs} performs canopy water interception and snow accumulation/melt.}
  \item{Function \code{hydrology_snowMelt} estimates snow melt using a simple energy balance, according to Kergoat (1998).}
}
}
\details{
The function simulates different vertical hydrological processes, which are described separately in other functions. 
If \code{modifySoil = TRUE} the function will modify the \code{soil} object (including both soil moisture and 
the snowpack on its surface) as a result of simulating hydrological processes.
}
\references{
Kergoat L. (1998). A model for hydrological equilibrium of leaf area index on a global scale. Journal of Hydrology 212–213: 268–286.
}
\seealso{
\code{\link{spwb_day}}, \code{\link{hydrology_rainInterception}}, \code{\link{hydrology_soilEvaporation}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
