% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{light_basic}
\alias{light_basic}
\alias{light_PARcohort}
\alias{light_PARground}
\alias{light_SWRground}
\alias{light_cohortAbsorbedSWRFraction}
\title{Basic radiation extinction}
\usage{
light_PARcohort(x, SpParams, gdd = NA_real_)

light_PARground(x, SpParams, gdd = NA_real_)

light_SWRground(x, SpParams, gdd = NA_real_)

light_cohortAbsorbedSWRFraction(z, x, SpParams, gdd = NA_real_)
}
\arguments{
\item{x}{An object of class \code{\link{forest}}}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}

\item{gdd}{Growth degree days.}

\item{z}{A numeric vector with height values.}
}
\description{
Basic radiation extinction
}
\seealso{
\code{\link{spwb}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
