% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mortality_dailyProbability}
\alias{mortality_dailyProbability}
\title{Mortality}
\usage{
mortality_dailyProbability(stressValue, stressThreshold)
}
\arguments{
\item{stressValue}{Current value of the stress variable (0 to 1, 
with higher values indicate stronger stress).}

\item{stressThreshold}{Threshold to indicate 50\% annual mortality probability.}
}
\value{
Returns a probability (between 0 and 1)
}
\description{
A simple sigmoid function to determine a daily mortality likelihood according 
to the value of a stress variable.
}
\seealso{
\code{\link{growth}}, \code{\link{regeneration}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
