% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.soil.R
\name{plot.soil}
\alias{plot.soil}
\title{Soil water curves}
\usage{
\method{plot}{soil}(
  x,
  type = "RetentionCurve",
  model = "SX",
  layer = 1,
  relative = TRUE,
  to = "SAT",
  ...
)
}
\arguments{
\item{x}{Initialized soil object (returned by function \code{\link{soil}}).}

\item{type}{A string of the plot type: "RetentionCurve" or "ConductivityCurve".}

\item{model}{model Either 'SX' or 'VG' for Saxton's or Van Genuchten's water retention models; or 'both' to plot both retention models.}

\item{layer}{Soil layer to be plotted.}

\item{relative}{Boolean flag to indicate that retention curve should be relative to field capacity or saturation.}

\item{to}{Either 'SAT' (saturation) or 'FC' (field capacity).}

\item{\dots}{Additional parameters to specific functions}
}
\value{
An object of class ggplot.
}
\description{
Function for plotting water retention or conductivity curves for soil layers.
}
\details{
The function is a common wrapper to specialized internal functions (see \code{\link{soil_retentionCurvePlot}}).
}
\examples{
# Initializes soil
s = soil(defaultSoilParams())
plot(s)

}
\seealso{
\code{\link{soil}}, \code{\link{summary.soil}}, \code{\link{soil_retentionCurvePlot}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
