% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_redefineLayers.R
\name{soil_redefineLayers}
\alias{soil_redefineLayers}
\title{Redefine soil layer widths}
\usage{
soil_redefineLayers(SoilParams, widths = c(300, 700, 1000, 2000))
}
\arguments{
\item{SoilParams}{A data frame of soil parameters (see an example in \code{\link{defaultSoilParams}}).}

\item{widths}{A numeric vector indicating the desired layer widths, in mm.}
}
\value{
A modified data frame of soil parameters
}
\description{
Allows redefining soil layer widths of an input data frame of soil parameters.
}
\seealso{
\code{\link{soil}}, \code{\link{defaultSoilParams}}
}
\author{
\enc{Víctor}{Victor} Granda, EMF-CREAF

Miquel De \enc{Cáceres}{Caceres} Ainsa, EMF-CREAF
}
