\encoding{UTF-8}
\docType{data}
\name{Parameter means}
\alias{trait_family_means}
\title{
Parameter average values
}
\description{
Internal data set with parameter averages for taxonomic families. This is used by input initialization functions to provide suitable parameter values when missing from species parameter tables.
}
\format{
Data frame \code{trait_family_means} has taxonomic families in rows and parameter names as columns.
}
\source{
Same sources as \code{\link{SpParamsMED}}
}
\seealso{\code{\link{SpParamsMED}}, \code{\link{spwbInput}}}
\keyword{datasets}
\examples{
medfate::trait_family_means
}
