% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_rockOptimization.R
\name{utils_rockOptimization}
\alias{utils_rockOptimization}
\title{Optimization of rock fragment content}
\usage{
utils_rockOptimization(
  x,
  soil,
  SpParams,
  control,
  meteo,
  PLCquantile = 0.9,
  qPLC_target = 12,
  qPLC_tol = 0.5,
  max_rocks = 99,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{An object of class \code{\link[medfate]{forest}}.}

\item{soil}{An object of class \code{\link{data.frame}} or \code{\link[medfate]{soil}}, containing soil parameters per soil layer.}

\item{SpParams}{A data frame with species parameters (see \code{\link[medfate]{SpParamsDefinition}} and \code{\link[medfate]{SpParamsMED}}).}

\item{control}{A list with default control parameters (see \code{\link[medfate]{defaultControl}}).}

\item{meteo}{A data frame with daily meteorological data series (see \code{\link[medfate]{spwb}}).}

\item{PLCquantile}{Maximum PLC quantile to be calculated across years.}

\item{qPLC_target}{Target PLC to be achieved (by default 12\%).}

\item{qPLC_tol}{Tolerance of PLC difference to target accepted when finding solution.}

\item{max_rocks}{Maximum content in coarse fragments allowed for any soil layer.}

\item{verbose}{A logical value. Print the internal messages of the function?}

\item{...}{Additional parameters to function \code{\link[medfate]{spwb}}.}
}
\value{
Function \code{utils_rockOptimization} returns a list containing the estimated rock fragment content,
the corresponding soil extractable water and the number of simulation runs performed.
}
\description{
Function \code{utils_rockOptimization} finds optimum rock fragment content in the soil
corresponding to given vegetation, weather and target percent loss
of conductance (PLC), following the method described in Druel et al. (2023).
}
\details{
The function performs a model inversion based on an ecohydrological assumption,
consisting in that forest leaf area index is in equilibrium with a low embolism
rate under normal conditions. This is translated in that the (by default 90\%) interannual quantile of
the maximum annual percent loss of conductance (PLC), averaged over plant cohorts,
should be close to a target PLC value (by default 12\%).
}
\examples{
\donttest{
#Load example daily meteorological data
data(examplemeteo)

#Load example plot plant data
data(exampleforest)

#Default species parameterization
data(SpParamsMED)

#Initialize soil with two layers
examplesoil <- defaultSoilParams(4)

#Rock fragment content optimization (Granier)
utils_rockOptimization(exampleforest, soil = examplesoil,
                       SpParams = SpParamsMED, meteo = examplemeteo,
                       control = defaultControl("Granier"),
                       elevation = 100, latitude = 41.82592)
}
}
\references{
Druel, A., Martins, N., Cochard, H., De Caceres, M., Delzon, S., Mencuccini, M., Torres-Ruiz, J., and Ruffault, J.: European forest vulnerability to hydraulic failure: an ecohydrological approach, EGU General Assembly 2023, Vienna, Austria, 24–28 Apr 2023, EGU23-17068, https://doi.org/10.5194/egusphere-egu23-17068, 2023.
}
\seealso{
\code{\link{spwb}}, \code{\link{soil}}, \code{\link{utils_ldrOptimization}}
}
\author{
\enc{Arsène}{Arsene} Druel, URFM-INRAE

Nicolas Martin-StPaul, URFM-INRAE

Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
