% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{communication}
\alias{communication}
\alias{.copy_model_output}
\title{Internal communication}
\usage{
.copy_model_output(internalCommunication, x, model)
}
\arguments{
\item{internalCommunication}{List for internal communication.}

\item{x}{An object of class \code{\link{spwbInput}} or \code{\link{growthInput}}.}

\item{model}{String for model, either "spwb" or "growth".}
}
\description{
Functions for internal communication. Not to be called by users.
}
\keyword{internal}
