% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomposition_DAYCENT}
\alias{decomposition_DAYCENT}
\title{DAYCENT decomposition}
\usage{
decomposition_DAYCENT(
  structuralLitter,
  CENTURYPools,
  paramsDecomposition,
  baseAnnualRates,
  annualTurnoverRate,
  sand,
  clay,
  soilTemperature,
  soilMoisture,
  soilPH,
  soilO2 = 1,
  cultfac = 1,
  tstep = 1
)
}
\arguments{
\item{structuralLitter}{A data frame with structural carbon pools corresponding to plant cohorts.}

\item{CENTURYPools}{A named numeric vector with metabolic, active, slow and passive carbon pools for surface and soil}

\item{soilTemperature}{Soil temperature (in Celsius).}

\item{soilMoisture}{Soil moisture content, relative to saturation.}

\item{soilPH}{Soil pH (0-14).}

\item{soilO2}{Soil oxygen factor.}

\item{cultfac}{Cultivation factor.}

\item{tstep}{Time step in days.}

\item{decompositionParams}{A data frame of species-specific decomposition parameters.}
}
\description{
DAYCENT decomposition
}
