% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decomposition_DAYCENTsnags}
\alias{decomposition_DAYCENTsnags}
\alias{decomposition_DAYCENTlitter}
\alias{decomposition_DAYCENT}
\title{DAYCENT decomposition}
\usage{
decomposition_DAYCENTsnags(
  snags,
  baseAnnualRates,
  airTemperature,
  airRelativeHumidity,
  tstep = 1
)

decomposition_DAYCENTlitter(
  litter,
  paramsLitterDecomposition,
  baseAnnualRates,
  sand,
  clay,
  soilTemperature,
  soilMoisture,
  soilPH,
  soilO2 = 1,
  cultfac = 1,
  tstep = 1
)

decomposition_DAYCENT(
  snags,
  litter,
  SOC,
  paramsLitterDecomposition,
  baseAnnualRates,
  annualTurnoverRate,
  airTemperature,
  airRelativeHumidity,
  sand,
  clay,
  soilTemperature,
  soilMoisture,
  soilPH,
  soilO2 = 1,
  cultfac = 1,
  tstep = 1
)
}
\arguments{
\item{snags}{A data frame with dead standing (snag) cohort information (see \code{\link{growthInput}}).}

\item{baseAnnualRates}{A named vector of annual decomposition rates, in yr-1 (see \code{\link{defaultControl}}).}

\item{airTemperature}{Mean daily air temperature (in Celsius).}

\item{airRelativeHumidity}{Mean daily relative humidity (\%).}

\item{tstep}{Time step in days. By default, one day. For annual time steps, use \code{tstep = 365.25}.}

\item{litter}{A data frame with aboveground and belowground structural carbon pools corresponding to plant cohorts, in g C/m2  (see \code{\link{growthInput}}).}

\item{paramsLitterDecomposition}{A data frame of species-specific litter decomposition parameters (see \code{\link{growthInput}}).}

\item{sand, clay}{Soil texture (sand and sand) in percent volume (\%).}

\item{soilTemperature}{Soil temperature (in Celsius).}

\item{soilMoisture}{Soil moisture content, relative to saturation (0-1).}

\item{soilPH}{Soil pH (0-14).}

\item{soilO2}{Soil oxygen factor (0-1).}

\item{cultfac}{Cultivation factor (0-1).}

\item{SOC}{A named numeric vector with metabolic, active, slow and passive carbon pools for surface and soil, in g C/m2  (see \code{\link{growthInput}}).}

\item{annualTurnoverRate}{Annual turnover rate, in yr-1  (see \code{\link{defaultControl}}).}
}
\value{
Function \code{decomposition_DAYCENTsnags} returns a vector containing transfer carbon flows to SOC pools and heterotrophic respiration from snag decomposition.
Function \code{decomposition_DAYCENTlitter} returns a vector containing transfer carbon flows to SOC pools and heterotrophic respiration from litter decomposition.
Function \code{decomposition_DAYCENT} returns scalar value with heterotrophic respiration (snags + litter + soil), in g C/m2.
}
\description{
Functions implementing a modification of the DAYCENT carbon decomposition model, inspired by the description given in Chapter 18 of Bonan (2019).
Functions \code{decompositionDAYCENTsnags} and \code{decompositionDAYCENTlitter} conduct snag and litter decomposition, respectively,
whereas function \code{decomposition_DAYCENT} performs the whole model for carbon decomposition.
}
\details{
Each call to functions \code{decomposition_DAYCENTlitter} or \code{decomposition_DAYCENTsnags} conducts one time step of the snag or litter dynamics, respectively.
Each call to function \code{decomposition_DAYCENT} conducts one time step of the whole DAYCENT
model and returns the heterotrophic respiration for that day.

\emph{IMPORTANT NOTE}: Decomposition functions modify the input data (i.e. \code{snags}, \code{litter} and/or \code{SOC}) according to decomposition rates and carbon transfer rates. When used as part of \code{\link{growth}} simulations,
soil physical and chemical parameters correspond to the uppermost soil layer.
}
\references{
Bonan, G. (2019). Climate change and terrestrial ecosystem modeling. Cambridge University Press, Cambridge, UK.
}
\seealso{
\code{\link{decomposition_temperatureEffect}}, \code{\link{growthInput}}, \code{\link{growth}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{internal}
