% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emptyforest.R
\name{emptyforest}
\alias{emptyforest}
\title{Creation of an empty forest}
\usage{
emptyforest(ntree = 0, nshrub = 0, nseed = 0)
}
\arguments{
\item{ntree, nshrub}{Number of tree and shrub cohorts, respectively.}

\item{nseed}{Number of species in the seed bank.}
}
\value{
An empty \code{\link{forest}} object.
}
\description{
Creates an empty \code{\link{forest}} object.
}
\examples{
# Initializes forest with 2 tree cohorts and 1 shrub cohort
emptyforest(ntree = 2, nshrub = 1)

}
\seealso{
\code{\link{forest}}, \code{\link{tree2forest}}, \code{\link{summary.forest}}, \code{\link{forest_mapWoodyTables}},  \code{\link{forest_mergeTrees}},  
\code{\link{plot.forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
