\encoding{UTF-8}
\name{exampleforest}
\Rdversion{1.1}
\alias{exampleforestMED}
\alias{exampleforestUS}
\docType{data}
\title{Example forest stands}
\description{
  Data set for illustration of model behaviour. Includes a description of the plant cohorts of a forest stand.
}
\usage{
data(exampleforestMED)
data(exampleforestUS)
}
\format{
 An object of class \code{\link{forest}} containing the description of the tree, sapling and shrub cohorts of a forest patch.
}
\source{
DGCN (2005). Tercer Inventario Forestal Nacional (1997-2007): Catalunya. \enc{Dirección}{Direccion} General de \enc{Conservación}{Conservacion} de la Naturaleza, Ministerio de Medio Ambiente, Madrid.
}
\seealso{ \code{\link{forest}}, \code{\link{spwb}}, \code{\link{forest2spwbInput}}}
\examples{
data(exampleforestMED)
}
\keyword{datasets}
