% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{herb_values}
\alias{herb_values}
\alias{herb_foliarBiomass}
\alias{herb_fuelLoading}
\alias{herb_LAI}
\title{Herbaceous description functions}
\usage{
herb_foliarBiomass(x, SpParams)

herb_fuelLoading(x, SpParams)

herb_LAI(x, SpParams)
}
\arguments{
\item{x}{An object of class \code{\link{forest}}.}

\item{SpParams}{A data frame with species parameters (see \code{\link{SpParamsMED}}).}
}
\value{
A single scalar:
\itemize{
\item{\code{herb_foliarBiomass}: Herbaceous biomass of leaves (in kg/m2).}
\item{\code{herb_fuelLoading}: Herbaceous fine fuel loading (in kg/m2).}
\item{\code{herb_LAI}: Herbaceous leaf area index (m2/m2).}
}
}
\description{
Functions to calculate attributes of the herbaceous component of a \code{\link{forest}} object
}
\seealso{
\code{\link{spwb}}, \code{\link{forest}}, \code{\link{plant_basalArea}}, \code{\link{summary.forest}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
\keyword{internal}
