% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hydrology_soilFlows}
\alias{hydrology_soilFlows}
\title{Soil flows}
\usage{
hydrology_soilFlows(
  soil,
  soilFunctions,
  rainfallInput,
  rainfallIntensity,
  snowmelt,
  sourceSink,
  infiltrationMode = "GreenAmpt1911",
  soilDomains = "single",
  freeDrainage = TRUE,
  nsteps = 24L,
  modifySoil = TRUE
)
}
\arguments{
\item{soil}{Object of class \code{\link{soil}}.}

\item{soilFunctions}{Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).}

\item{rainfallInput}{Amount of water from rainfall event (after excluding interception), in mm.}

\item{rainfallIntensity}{Rainfall intensity, in mm/h.}

\item{snowmelt}{Amount of water originated from snow melt, in mm.}

\item{sourceSink}{Source/sink term for each soil layer (from soil evaporation or plant transpiration/redistribution)
as mm/day.}

\item{infiltrationMode}{Infiltration model, either "GreenAmpt1911" or "Boughton1989"}

\item{soilDomains}{Either "single" (for single-domain) or "dual" (for dual-permeability).}

\item{freeDrainage}{Boolean flag to indicate that lower boundary condition is free drainage.}

\item{nsteps}{Number of time steps per day}

\item{modifySoil}{Boolean flag to indicate that the input \code{soil} object should be modified during the simulation.}
}
\value{
Returns a named vector with three elements:
  \itemize{
    \item{\code{Infiltration}: water infiltrated into the soil.}
    \item{\code{Runoff}: surface runoff generated.}
    \item{\code{DeepDrainage}: the water draining from the bottom layer.}
  }
}
\description{
Function \code{hydrology_soilFlows} estimates water movement within the soil according to Richards equation.
}
\seealso{
\code{\link{spwb}}, \code{\link{hydrology_soilWaterInputs}}, \code{\link{hydrology_infiltration}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
