% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{hydrology_soilWaterBalance}
\alias{hydrology_soilWaterBalance}
\title{Soil flows}
\usage{
hydrology_soilWaterBalance(
  soil,
  soilFunctions,
  rainfallInput,
  rainfallIntensity,
  snowmelt,
  sourceSink,
  infiltrationMode = "GreenAmpt1911",
  soilDomains = "single",
  freeDrainage = TRUE,
  nsteps = 24L,
  modifySoil = TRUE
)
}
\arguments{
\item{soil}{Object of class \code{\link{soil}}.}

\item{soilFunctions}{Soil water retention curve and conductivity functions, either 'SX' (for Saxton) or 'VG' (for Van Genuchten).}

\item{rainfallInput}{Amount of water from rainfall event (after excluding interception), in mm.}

\item{rainfallIntensity}{Rainfall intensity, in mm/h.}

\item{snowmelt}{Amount of water originated from snow melt, in mm.}

\item{sourceSink}{Source/sink term for each soil layer (from soil evaporation or plant transpiration/redistribution)
as mm/day.}

\item{infiltrationMode}{Infiltration model, either "GreenAmpt1911" or "Boughton1989"}

\item{soilDomains}{Either "single" (for single-domain) or "dual" (for dual-permeability).}

\item{freeDrainage}{Boolean flag to indicate that lower boundary condition is free drainage.}

\item{nsteps}{Number of time steps per day}

\item{modifySoil}{Boolean flag to indicate that the input \code{soil} object should be modified during the simulation.}
}
\value{
Returns a named vector with different elements, depending on \code{soilDomains}. If
  \code{soilDomains == "single"}:
  \itemize{
    \item{\code{Snowmelt}: Snowmelt input (mm).}
    \item{\code{Source/sinks}: Sum of source/sink input across layers (mm).}
    \item{\code{Infiltration}: Water infiltrated into the soil (mm).}
    \item{\code{Runoff}: Surface runoff generated (mm).}
    \item{\code{DeepDrainage}: Water draining from the bottom layer (mm). This quantity is corrected to close the water balance.}
    \item{\code{VolumeChange}: Change in soil water volume (mm).}
    \item{\code{UncorrectedWaterBalance}: Uncorrected balance of inputs/outputs (mm).}
  }
 If \code{soilDomains == "dual"} the named vector contains the following additional elements:
  \itemize{
    \item{\code{Lateral flows}: Sum of water circulating between micropores and macropores, positive when filling micropores (mm).}
    \item{\code{InfiltrationMicropores}: Water infiltrated into the soil matrix (mm).}
    \item{\code{InfiltrationMacropores}: Water infiltrated into the soil macropore domain (mm).}
    \item{\code{DrainageMicropores}: Water draining from the bottom layer of the micropore domain (mm). This quantity is corrected to close water balance in the micropore domain.}
    \item{\code{DrainageMacropores}: Water draining from the bottom layer of the macropore domain (mm). This quantity is corrected to close the water balance in the macropore domain.}
    \item{\code{MicroporeVolumeChange}: Change in soil water volume in the micropore domain (mm).}
    \item{\code{UncorrectedMicroporeBalance}: Uncorrected balance of inputs/outputs in the micropore domain (mm).}
    \item{\code{MacroporeVolumeChange}: Change in soil water volume in the macropore domain (mm).}
    \item{\code{UncorrectedMacroporeBalance}: Uncorrected balance of inputs/outputs in the macropore domain (mm).}
  }
}
\description{
Function \code{hydrology_soilFlows} estimates water balance of soil layers given water inputs/outputs, including the simulation of water movement within the soil.
}
\examples{
# Initialize soil example
examplesoil <- soil(defaultSoilParams(4))

# Water balance in a single-domain simulation (Richards equation)
hydrology_soilWaterBalance(examplesoil, "VG", 10, 5, 0, c(-1,-1,-1,-1), 
                           soilDomains = "single", modifySoil = FALSE)
                    
# Water balance in a dual-permeability model (MACRO)
hydrology_soilWaterBalance(examplesoil, "VG", 10, 5, 0, c(-1,-1,-1,-1), 
                           soilDomains = "dual", modifySoil = FALSE)
  
}
\seealso{
\code{\link{spwb}}, \code{\link{hydrology_soilWaterInputs}}, \code{\link{hydrology_infiltration}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF

\enc{María González Sanchís}{Maria Gonzalez Sanchis}, UPV-CTFC
}
