% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{resetInputs}
\alias{resetInputs}
\title{Reset simulation inputs}
\usage{
resetInputs(x)
}
\arguments{
\item{x}{An object of class \code{\link{spwbInput}} or \code{\link{growthInput}}.}
}
\value{
Does not return any value. Instead, it modifies input object \code{x}.
}
\description{
Function \code{resetInputs()} allows resetting state variables in \code{x} to their defaults.
}
\seealso{
\code{\link{spwbInput}}, \code{\link{growthInput}}, \code{\link{spwb}}
}
\author{
Miquel De \enc{Cáceres}{Caceres} Ainsa, CREAF
}
