% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{SpParams}
\alias{SpParams}
\alias{SpParamsDefinition}
\alias{SpParamsMED}
\alias{SpParamsES}
\alias{SpParamsFR}
\alias{SpParamsUS}
\title{Data tables with species parameter definition and values for different countries}
\format{
\itemize{
  \item{Data frame \code{SpParamsDefinition} has parameters in rows and columns 'ParameterName', 'ParameterGroup', 'Definition', 'Type' and 'Units'.}
  \item{Data frames \code{SpParamsMED} (for Catalonia), \code{SpParamsES} (for Spain), \code{SpParamsFR} (for France) and \code{SpParamsUS} (for US) have species or genus as rows and column names equal to parameter names in \code{SpParamsDefinition}.}
}
}
\description{
A data sets of species parameter definition and values, the latter resulting from existing databases, fit to empirical data or expert-based guesses.
}
\details{
Plant trait parameter sources are listed in the bibliography section. Details of the procedures used to obtain the species parameter tables can be found in an article at https://emf-creaf.github.io/medfate/.
}
\examples{
data(SpParamsDefinition)
data(SpParamsMED)
}
\keyword{data}
